requirements.txt
django
konlpy
drf-yasg
djangorestframework

